fs: rewrite BiDiStream's client->server implementation.
Prevously every reader got its own channel to the Stream
for writing. This ended up being gross, because reflection
was required to select all the channels, new channels could
be added at any time, requiring signalling into Read() to
tell it to loop around the select when new channels were added,
and other hacky stuff. This is replaced by an implementation
where each reader gets a reference to the same Stream channel.
That channel is never closed, since it is not simple to do safely.
Instead, we give each reader a handle to the Stream's new 'close'
channel, which is closed when the stream is closed, so readers
calling Write() will check this channel and return EOF if the
stream is closed.

There *is* still a race condition, but it requires 3 goroutines to
all be perfectly aligned to hit an unlikely case. The race will cause
a Write() payload to be dropped. This is acceptable for now.