fs: add PipeFile, add stream documentation, EOF and other fixes
This adds PipeFile, which is a fs.File that behaves much like a network
socket. Each Open() of a PipeFile is given its own stream and handled in
its own goroutine. This allows clients to communicate with the server
independently of one another. This is useful for serving protocols.

This also adds some basic documentation for the streams.

Previously, a client reading a stream which had been closed would receive
an error (9p rerror) with the EOF message. This was not correct. Now, reads
at the end of the closed stream return rread with zero length, as they
should.

This also fixes a panic caused by closing a stream more than once.